+++ api/Model/Paiement.cs
@@ -12,13 +12,13 @@ namespace api.Model
         public int Id { get; set; }
         public int Annee { get; set; }
         public int Mois { get; set; }
-        public float SalaireDeBase { get; set; }
-        public float Prime { get; set; }
+        public double SalaireDeBase { get; set; }
+        public double Prime { get; set; }
         public int Nmbrheursupplimentaires { get; set; }
         public int NmbrAbscences { get; set; }
-        public float ImpotSurSalaire { get; set; }
-        public float CNSS { get; set; }
-        public float AMO { get; set; }
+        public double ImpotSurSalaire { get; set; }
+        public double CNSS { get; set; }
+        public double AMO { get; set; }
         public string FicheDePaie { get; set; } = "";
         public string? AppUserId { get; set; }
         public AppUser? AppUser { get; set; }
